HSEARCH-3107 Remove old (stateful) DomainDataUpdater
It is not used anymore
